requirements.txt
discord.py
pytz.py